AK: Move AK::Detail::StringData to its own header file
This will allow us to access it from FlyString.cpp
